Added Proxy examples
Added virtual proxy package that contained a remote example.  Keep in mind
that this won't work out of the box - it required compiling with javac,
and manually running rmiregistry, which is rightfully deprecated because
it sucks.  I also added the virtual proxy package which I will likely use
in Whackscend